Move CLI time utilities into shared boundary
Relocated the general time helpers from src/cli/features/shared/time-utils.js to src/cli/shared/time-utils.js so the shared directory hosts the CLI-wide utilities while the feature shim re-exports them for compatibility. This keeps feature modules focused on feature-specific logic and makes imports match the shared dependency surface.
