+++ src/cli/commands/generate-feather-metadata.js
@@ -11,10 +11,7 @@ import {
 } from "../shared/dependencies.js";
 import { CliUsageError } from "../core/errors.js";
 import { assertSupportedNodeVersion } from "../shared/node-version.js";
-import {
-    timeSync,
-    createVerboseDurationLogger
-} from "../features/shared/time-utils.js";
+import { timeSync, createVerboseDurationLogger } from "../shared/time-utils.js";
 import {
     disposeProgressBars,
     withProgressBarCleanup
