+++ src/cli/commands/generate-gml-identifiers.js
@@ -17,10 +17,7 @@ import {
     createManualDownloadReporter,
     downloadManualFileEntries
 } from "../features/manual/utils.js";
-import {
-    timeSync,
-    createVerboseDurationLogger
-} from "../features/shared/time-utils.js";
+import { timeSync, createVerboseDurationLogger } from "../shared/time-utils.js";
 import {
     disposeProgressBars,
     withProgressBarCleanup
